+++ examples/apps/ecto_example/lib/ecto_example/postgres_repo.ex
@@ -1,0 +1,5 @@
+defmodule EctoExample.PostgresRepo do
+  use Ecto.Repo,
+    otp_app: :ecto_example,
+    adapter: Ecto.Adapters.Postgres
+end