+++ lib/new_relic/config.ex
@@ -92,12 +92,29 @@ defmodule NewRelic.Config do
   Some Agent features can be controlled via configuration
 
   * `:error_collector_enabled` (default `true`)
+    * Controls collecting any Error traces or metrics
+  * `:sql_collection_enabled` (default `true`)
+    * Controls collection of SQL query strings
+  * `:ecto_instrumentation_enabled` (default `true`)
+    * Controls all Ecto instrumentation
   """
   def feature?(:error_collector) do
-    case System.get_env("NEW_RELIC_ERROR_COLLECTOR_ENABLED") do
+    feature_check?("NEW_RELIC_ERROR_COLLECTOR_ENABLED", :error_collector_enabled)
+  end
+
+  def feature?(:sql_collection) do
+    feature_check?("NEW_RELIC_SQL_COLLECTION_ENABLED", :sql_collection_enabled)
+  end
+
+  def feature?(:ecto_instrumentation) do
+    feature_check?("NEW_RELIC_ECTO_INSTRUMENTTION_ENABLED", :ecto_instrumentation_enabled)
+  end
+
+  defp feature_check?(env, config) do
+    case System.get_env(env) do
       "true" -> true
       "false" -> false
-      _ -> Application.get_env(:new_relic_agent, :error_collector_enabled, true)
+      _ -> Application.get_env(:new_relic_agent, config, true)
     end
   end